refactor, clippy, fmt
