+++ prql-compiler/src/parser/expr.rs
@@ -1,167 +1,11 @@
-use std::{collections::HashMap, str::FromStr};
+use std::collections::HashMap;
 
 use chumsky::prelude::*;
 use itertools::Itertools;
-use semver::VersionReq;
 
+use super::common::*;
 use super::lexer::{InterpolItem, Token};
-use crate::{ast::pl::*, Span};
-
-pub fn source() -> impl Parser<Token, Vec<Stmt>, Error = Simple<Token>> {
-    query_def()
-        .or_not()
-        .chain(
-            var_def()
-                .or(function_def())
-                .separated_by(new_line().or(whitespace()).repeated()),
-        )
-        .chain(main_pipeline().or_not())
-        .then_ignore(end())
-        .labelled("source file")
-}
-
-fn main_pipeline() -> impl Parser<Token, Stmt, Error = Simple<Token>> {
-    pipeline(expr_call())
-        .map_with_span(into_expr)
-        .map(Box::new)
-        .map(StmtKind::Main)
-        .map_with_span(into_stmt)
-        .labelled("main pipeline")
-}
-
-fn query_def() -> impl Parser<Token, Stmt, Error = Simple<Token>> {
-    keyword("prql")
-        .ignore_then(
-            // named arg
-            ident_part()
-                .then_ignore(ctrl(":").padded_by(whitespace().or_not()))
-                .then(expr_call())
-                .repeated(),
-        )
-        .then_ignore(whitespace().or_not().then(new_line()))
-        .try_map(|args, span| {
-            let mut params: HashMap<_, _> = args.into_iter().collect();
-
-            let version = params
-                .remove("version")
-                .map(|v| match v.kind {
-                    ExprKind::Literal(Literal::String(v)) => {
-                        VersionReq::parse(&v).map_err(|e| e.to_string())
-                    }
-                    _ => Err("version must be a sting literal".to_string()),
-                })
-                .transpose()
-                .map_err(|msg| Simple::custom(span, msg))?;
-
-            let other = params
-                .into_iter()
-                .flat_map(|(key, value)| match value.kind {
-                    ExprKind::Ident(value) => Some((key, value.to_string())),
-                    _ => None,
-                })
-                .collect();
-
-            Ok(StmtKind::QueryDef(QueryDef { version, other }))
-        })
-        .map_with_span(into_stmt)
-        .labelled("query header")
-}
-
-fn var_def() -> impl Parser<Token, Stmt, Error = Simple<Token>> {
-    keyword("let")
-        .ignore_then(whitespace())
-        .ignore_then(ident_part())
-        .then_ignore(ctrl("=").padded_by(whitespace().or_not()))
-        .then(expr_call().map(Box::new))
-        .map(|(name, value)| VarDef { name, value })
-        .map(StmtKind::VarDef)
-        .map_with_span(into_stmt)
-        .labelled("variable definition")
-}
-
-fn function_def() -> impl Parser<Token, Stmt, Error = Simple<Token>> {
-    keyword("func")
-        .ignore_then(whitespace())
-        .ignore_then(
-            // func name
-            ident_part()
-                .then_ignore(whitespace().or_not())
-                .then(type_expr().or_not()),
-        )
-        .then(
-            // params
-            ident_part()
-                .then_ignore(whitespace().or_not())
-                .then(type_expr().or_not())
-                .then_ignore(whitespace().or_not())
-                .then(
-                    ctrl(":")
-                        .ignore_then(whitespace().or_not())
-                        .ignore_then(expr())
-                        .or_not(),
-                )
-                .repeated(),
-        )
-        .then_ignore(whitespace().or_not())
-        .then_ignore(ctrl("->"))
-        .then(expr_call().map(Box::new))
-        .then_ignore(whitespace().or_not())
-        .then_ignore(new_line())
-        .map(|(((name, return_ty), params), body)| {
-            let (pos, nam) = params
-                .into_iter()
-                .map(|((name, ty), default_value)| FuncParam {
-                    name,
-                    ty,
-                    default_value,
-                })
-                .partition(|p| p.default_value.is_none());
-
-            FuncDef {
-                name,
-                positional_params: pos,
-                named_params: nam,
-                body,
-                return_ty,
-            }
-        })
-        .map(StmtKind::FuncDef)
-        .map_with_span(into_stmt)
-        .labelled("function definition")
-}
-
-pub fn type_expr() -> impl Parser<Token, Ty, Error = Simple<Token>> {
-    recursive(|type_expr| {
-        let type_term = ident_part().then(type_expr.or_not()).map(|(name, param)| {
-            let ty = match TyLit::from_str(&name) {
-                Ok(t) => Ty::from(t),
-                Err(_) if name == "table" => Ty::Table(Frame::default()),
-                Err(_) => {
-                    eprintln!("named type: {}", name);
-                    Ty::Named(name.to_string())
-                }
-            };
-
-            if let Some(param) = param {
-                Ty::Parameterized(Box::new(ty), Box::new(param))
-            } else {
-                ty
-            }
-        });
-
-        ctrl("<")
-            .ignore_then(type_term.separated_by(ctrl("|").padded_by(whitespace().or_not())))
-            .then_ignore(ctrl(">"))
-            .map(|mut terms| {
-                if terms.len() == 1 {
-                    terms.remove(0)
-                } else {
-                    Ty::AnyOf(terms)
-                }
-            })
-    })
-    .labelled("type expression")
-}
+use crate::ast::pl::*;
 
 pub fn expr_call() -> impl Parser<Token, Expr, Error = Simple<Token>> {
     func_call(expr())
@@ -247,7 +91,7 @@ pub fn expr() -> impl Parser<Token, Expr, Error = Simple<Token>> + Clone {
     })
 }
 
-fn pipeline<E>(expr: E) -> impl Parser<Token, ExprKind, Error = Simple<Token>>
+pub fn pipeline<E>(expr: E) -> impl Parser<Token, ExprKind, Error = Simple<Token>>
 where
     E: Parser<Token, Expr, Error = Simple<Token>>,
 {
@@ -314,7 +158,7 @@ where
             alias: Some(alias),
             ..expr
         });
-    let positional_arg = assign_call.or(expr.clone()).map(|expr| (None, expr));
+    let positional_arg = assign_call.or(expr).map(|expr| (None, expr));
 
     let args = whitespace()
         .ignore_then(named_arg.or(positional_arg))
@@ -386,45 +230,4 @@ fn operator_logical() -> impl Parser<Token, BinOp, Error = Simple<Token>> {
 }
 fn operator_coalesce() -> impl Parser<Token, BinOp, Error = Simple<Token>> {
     ctrl("??").to(BinOp::Coalesce)
-}
-
-fn ident_part() -> impl Parser<Token, String, Error = Simple<Token>> {
-    select! { Token::Ident(ident) => ident }
-}
-
-fn keyword(kw: &str) -> impl Parser<Token, (), Error = Simple<Token>> + '_ {
-    select! { Token::Ident(ident) if ident == kw => () }
-}
-
-fn whitespace() -> impl Parser<Token, (), Error = Simple<Token>> + Clone {
-    just(Token::Whitespace).repeated().at_least(1).to(())
-}
-
-fn new_line() -> impl Parser<Token, (), Error = Simple<Token>> + Clone {
-    just(Token::NewLine).to(())
-}
-
-fn ctrl(chars: &'static str) -> impl Parser<Token, (), Error = Simple<Token>> {
-    select! { Token::Control(str) if str == chars => () }
-}
-
-fn into_stmt(kind: StmtKind, span: std::ops::Range<usize>) -> Stmt {
-    Stmt {
-        span: into_span(span),
-        ..Stmt::from(kind)
-    }
-}
-
-fn into_expr(kind: ExprKind, span: std::ops::Range<usize>) -> Expr {
-    Expr {
-        span: into_span(span),
-        ..Expr::from(kind)
-    }
-}
-
-fn into_span(span: std::ops::Range<usize>) -> Option<Span> {
-    Some(Span {
-        start: span.start,
-        end: span.end,
-    })
 }
